requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
arq
redis
google-generativeai
python-dotenv
supabase
jobspy
pandas
linkedin-jobs-scraper